refactor: LottoManager의 기능을 LottoMachine에 위임
- 기존에는 현실 세계를 반영하여 LottoManger를 두었다.
하지만 LottoManager에 별다른 로직이 존재하지 않기도 하고,
LottoMachine이 구입금액에 따른 로또들을 발행하는 것도 적절하다고 판단됐다.
